prevent player from "sticking" to walls
previously, when you would "push against a wall" in midair, the player
wouldn't fall (because of the friction). we fix this by making the
player frictionless (which obviously then causes them to slide along
floors). to fix the sliding, manually dampen the horizontal movement
each frame when not pressing a movement key